Added :Ability to throw User Friendly Errors in case Employee Id is not found in Employee Payroll App UC12.

+++ src/main/java/com/bridgelabz/employeepayrollservice/service/EmployeePayrollService.java
@@ -1,31 +1,24 @@
 package com.bridgelabz.employeepayrollservice.service;
 
-import java.util.List;
-import java.util.Optional;
-
 import com.bridgelabz.employeepayrollservice.dto.EmployeeDTO;
+import com.bridgelabz.employeepayrollservice.exception.EmployeePayrollException;
 import com.bridgelabz.employeepayrollservice.model.Employee;
 import com.bridgelabz.employeepayrollservice.repository.EmployeePayrollRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Optional;
+
 
 @Service
+
+//Created EmployeePayrollService class to serve api calls done by controller layer
 public class EmployeePayrollService implements IEmployeePayrollService {
+
+    //Autowired EmployeePayrollRepository interface to inject its dependency here
     @Autowired
     EmployeePayrollRepository repository;
-
-    public String getMessage(String name) {
-        return "Welcome " + name;
-    }
-
-    public String postMessage(EmployeeDTO employeeDTO) {
-        return "Hello " + employeeDTO.getFirstName() + "" + employeeDTO.getLastName() + "!";
-    }
-
-    public String putMessage(String name) {
-        return "How are you, " + name;
-    }
 
     public String getWelcome() {
         return "Welcome to Employee Payroll !!!";
@@ -42,19 +35,36 @@ public class EmployeePayrollService implements IEmployeePayrollService {
         return list;
     }
 
-    public Optional<Employee> getDataById(Integer id) {
-        Employee newEmployee = repository.getById(id);
-        return Optional.of(newEmployee);
+    public Employee getDataById(Integer id) {
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            return newEmployee.get();
+        } else throw new EmployeePayrollException("Employee id not found");
     }
 
     public Employee updateDataById(Integer id, EmployeeDTO employeeDTO) {
-        Employee newEmployee = new Employee(id, employeeDTO);
-        repository.save(newEmployee);
-        return newEmployee;
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            Employee employee = new Employee(id, employeeDTO);
+            repository.save(employee);
+            return employee;
+        } else {
+            throw new EmployeePayrollException("Employee Not found");
+        }
     }
 
     public String deleteDataById(Integer id) {
-        repository.deleteById(id);
-        return null;
+        Optional<Employee> newEmployee = repository.findById(id);
+        if (newEmployee.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new EmployeePayrollException("Employee Details not found");
+        }
+      return null;
     }
 }
+
+
+
+
+
